+++ webviz_config/_localhost_token.py
@@ -7,14 +7,13 @@ from ._is_reload_process import is_reload_process
 
 
 class LocalhostToken:
-    """Uses a method similar to jupyter notebook (however, here we do it over
-    https in addition). This method is only used during interactive usage on
-    localhost, and the workflow is as follows:
+    """Uses a method similar to jupyter notebook. This method is only used during
+    interactive usage on localhost, and the workflow is as follows:
 
     - During the flask app building, a one-time-token (ott) and a cookie_token
       is generated.
     - When the app is ready, the user needs to "login" using this
-      one-time-token in the url (https://localhost:{port}?ott={token})
+      one-time-token in the url (http://localhost:{port}?ott={token})
     - If ott is valid - a cookie with a separate token is set, and the
       one-time-token is discarded. The cookie is then used for subsequent
       requests.